add tldrw file

+++ headers/fractol.h
@@ -6,7 +6,7 @@
 /*   By: baouragh <baouragh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/05 09:24:48 by baouragh          #+#    #+#             */
-/*   Updated: 2024/02/28 18:22:22 by baouragh         ###   ########.fr       */
+/*   Updated: 2024/03/04 10:39:25 by baouragh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,14 +87,14 @@ typedef struct s_fractal
     
     int iter;
     int color;
-    int x;
-    int y;
+    double x;
+    double y;
     int check_set;
     int argc;
     int max_iter;
     
-    int mouse_x;
-    int mouse_y;
+    double mouse_x;
+    double mouse_y;
     double x_tmp;
     double x_shift_value;
     double y_shift_value;
